runtimes/eoapi/raster/pyproject.toml
[project]
name = "eoapi.raster"
description = "Raster service for eoAPI."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
]
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: GIS",
]
dynamic = ["version"]
dependencies = [
    "titiler.pgstac==1.7.1",
    "titiler.extensions",
    "starlette-cramjam>=0.4,<0.5",
    "importlib_resources>=1.1.0;python_version<'3.9'",
    "boto3",
    "eoapi.auth-utils>=0.2.0",
]

[project.optional-dependencies]
# https://www.psycopg.org/psycopg3/docs/api/pq.html#pq-module-implementations
psycopg = [  # pure python implementation
    "psycopg[pool]"
]
psycopg-c = [  # C implementation of the libpq wrapper
    "psycopg[c,pool]"
]

psycopg-binary = [  # pre-compiled C implementation
    "psycopg[binary,pool]"
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.pdm.version]
source = "file"
path = "eoapi/raster/__init__.py"

[tool.pdm.build]
includes = ["eoapi/raster"]
excludes = ["tests/", "**/.mypy_cache", "**/.DS_Store"]
